Extract interface MatchResult
Convert MatchResult to an interface and move the data-verification-
specific implementation of the MatchResult to DataMatchResult.
