+++ presto-verifier/src/test/java/com/facebook/presto/verifier/resolver/FailureResolverTestUtil.java
@@ -19,12 +19,12 @@ import com.facebook.presto.sql.tree.Identifier;
 import com.facebook.presto.verifier.checksum.ColumnChecksum;
 import com.facebook.presto.verifier.checksum.ColumnMatchResult;
 import com.facebook.presto.verifier.framework.Column;
-import com.facebook.presto.verifier.framework.MatchResult;
+import com.facebook.presto.verifier.framework.DataMatchResult;
 
 import java.util.Optional;
 import java.util.OptionalLong;
 
-import static com.facebook.presto.verifier.framework.MatchResult.MatchType.COLUMN_MISMATCH;
+import static com.facebook.presto.verifier.framework.DataMatchResult.MatchType.COLUMN_MISMATCH;
 import static java.util.Arrays.asList;
 
 public class FailureResolverTestUtil
@@ -37,9 +37,9 @@ public class FailureResolverTestUtil
         return new ColumnMatchResult<>(false, column, controlChecksum, testChecksum);
     }
 
-    public static MatchResult createMatchResult(ColumnMatchResult<?>... mismatchedColumns)
+    public static DataMatchResult createMatchResult(ColumnMatchResult<?>... mismatchedColumns)
     {
-        return new MatchResult(COLUMN_MISMATCH, Optional.empty(), OptionalLong.of(1L), OptionalLong.of(1L), asList(mismatchedColumns));
+        return new DataMatchResult(COLUMN_MISMATCH, Optional.empty(), OptionalLong.of(1L), OptionalLong.of(1L), asList(mismatchedColumns));
     }
 
     public static SqlVarbinary binary(int data)
